+++ src/index.js
@@ -14,18 +14,28 @@ import { hasText, resolveSecretsName } from '@adobe/spacecat-shared-utils';
 import { badRequest, internalServerError, notFound } from '@adobe/spacecat-shared-http-utils';
 import { helixStatus } from '@adobe/helix-status';
 import secrets from '@adobe/helix-shared-secrets';
+import dataAccess from '@adobe/spacecat-shared-data-access';
 import apex from './apex/handler.js';
 import cwv from './cwv/handler.js';
 import sitemap from './sitemap/handler.js';
-import notFoundHandler from './notfound/handler.js';
-import backlinks from './backlinks/handler.js';
+import noopHandler from './digest/handler-noop.js';
+import notFoundInternalDigestHandler from './notfound/handler-internal.js';
+import notFoundExternalDigestHandler from './notfound/handler-external.js';
+import brokenBacklinksInternal from './backlinks/handler-internal.js';
+import brokenBacklinksExternal from './backlinks/handler-external.js';
+import experimentation from './experimentation/handler.js';
 
 export const HANDLERS = {
   apex,
   cwv,
   sitemap,
-  404: notFoundHandler,
-  'broken-backlinks': backlinks,
+  404: noopHandler,
+  '404-external': notFoundExternalDigestHandler,
+  '404-internal': notFoundInternalDigestHandler,
+  'broken-backlinks': noopHandler,
+  'broken-backlinks-external': brokenBacklinksExternal,
+  'broken-backlinks-internal': brokenBacklinksInternal,
+  experimentation,
 };
 
 function guardEnvironmentVariables(fn) {
@@ -74,7 +84,9 @@ async function run(message, context) {
     url,
   } = message;
 
-  log.info(`Audit result received for url: ${url}\nmessage content: ${JSON.stringify(message)}`);
+  if (url) {
+    log.info(`Audit result received for url: ${url}\nmessage content: ${JSON.stringify(message)}`);
+  }
 
   const handler = HANDLERS[type];
   if (!handler) {
@@ -95,6 +107,7 @@ async function run(message, context) {
 }
 
 export const main = wrap(run)
+  .with(dataAccess)
   .with(sqsEventAdapter)
   .with(guardEnvironmentVariables)
   .with(secrets, { name: resolveSecretsName })